feat(examples): fold the Live ticker page into the Lifecycle guide (phase 3)
Eleventh phase-3 cluster (Components, part 2b). The standalone /realtime/{Symbol}
page is removed; its reusable LiveTicker widget (poll loop in OnMountAsync, a
symbol switch that fires OnPropsChanged*, OnRendered first-paint, OnUnmount*,
CancellationToken, and a zero-JS server-rendered SVG chart) is embedded in
docs/lifecycle.md under 'When OnPropsChanged* refires' via a new LiveTickerDemo
(registered lifecycle-ticker).

A co-mounted guide demo can't own a live [RouteParam], so the BTC/ETH/SOL switcher
now flips Symbol via internal state instead of nav.NavigateTo — re-rendering
reconciles the same LiveTicker instance at its stable position and fires
OnPropsChanged exactly as the route-param switch did.

- LiveTicker.cs / PricePoint.cs unchanged; new LiveTickerDemo.cs wraps them.
- LiveTickerPageTests -> LiveTickerDemoTests (render the demo directly); the
  obsolete MatchPrefix render-through-App test is removed (its logic stays covered
  by IsActive_HrefWithMatchPrefix); PageBaselineTests rows dropped.
- Sidebar row + home card repoint at the guide; E2E ticker walk moves to
  WalkLifecycleGuideAsync (URL assertion dropped, internal switch now).

TablePage, OrdersPage and the Todos capstone are untouched (final cluster).
